refactor: merge transpose partial encoder/decoder into `TransposeCodecPartial`
